// chpqa-app/src/app/api-services/chpqa-api/generated/model/requestUpdateSubmissionHeatOutput.ts
/**
 * WebApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RequestHeatOutput } from './requestHeatOutput';


export interface RequestUpdateSubmissionHeatOutput { 
    idSubmission: string;
    qualifyingHeatOutput: number;
    totalHeatExported?: number | null;
    estimatedTotalHeatOutputUsedInthePrimeMovers?: number | null;
    estimatedTotalHeatOutputUsedIntheBoilers?: number | null;
    heatOutputList: Array<RequestHeatOutput>;
}